프로젝트의 페이지 전용 컴포트와 중복된 기능을 사용하는 컴포넌트로 폴더를 나눕니다.
