+++ src/components/Header.js
@@ -2,6 +2,7 @@ import React from "react";
 import profile from "../assets/profile.png";
 import search from "../assets/search.PNG";
 import logo from "../assets/logo.png";
+import { Link } from "react-router-dom";
 
 import "./Header.css";
 
@@ -22,10 +23,16 @@ const Header = () => {
               <div id="header-info-content">
                 <span id="fb-info">
                   <img className="header-profile" src={profile} alt={profile} />
-                  <span>김우석</span>
+                  <Link to="Private">
+                    <span className="Header-name">김우석</span>
+                  </Link>
                 </span>
                 <span class="fb-spliter"> | </span>
-                <span id="fb-home">홈</span>
+                <Link to="/">
+                  <span className="Header-home" id="fb-home">
+                    홈
+                  </span>
+                </Link>
               </div>
             </div>
           </div>
