+++ src/components/Myheader.css
@@ -124,3 +124,7 @@
     color: #365899;
     text-decoration: none
 }
+
+.navigation-content {
+    color: #365899;
+}
