+++ src/components/Myheader.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { Link } from "react-router-dom";
 import "./Myheader.css";
 
 import profile from "../assets/profile.png";
@@ -22,23 +23,28 @@ const Myheader = () => {
           <button className="profile-save">프로필 저장</button>
           <img src={profile} class="photo" alt={profile}></img>
           <img src={background} class="top" alt={background}></img>
-          <p class="button1">프로필 수정</p>
+          <Link to="ProfileEdit">
+            <p class="button1">프로필 수정</p>
+          </Link>
+
           <div class="bottom">
             <ul class="navigation">
               <li>
-                <div href="">타임라인</div>
+                <div className="navigation-content">타임라인</div>
               </li>
               <li>
-                <div href="">정보</div>
+                <div className="navigation-content">정보</div>
               </li>
               <li>
-                <div href="">친구</div>
+                <div className="navigation-content">친구</div>
               </li>
               <li>
-                <div href="">사진</div>
+                <Link to="Picture">
+                  <div className="navigation-content">사진</div>
+                </Link>
               </li>
               <li>
-                <div href="">더보기</div>
+                <div className="navigation-content">더보기</div>
               </li>
             </ul>
           </div>
