--- src/components/Picturepage.css
@@ -1,97 +1,0 @@
-*,
-
-html {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
-
-header,
-nav,
-footer,
-section {
-    display: block;
-}
-
-body {
-    background-color: #E9EBEE;
-    font-family: serif, sans-serif;
-    min-width: 1340px;
-}
-
-.middle {
-    width: 1050px;
-    margin: auto auto;
-}
-
-.right {
-    padding-left: 100px
-}
-
-.profile-pic {
-    width: 170px;
-    height: 170px;
-    border: 5px solid white;
-    position: absolute;
-    background: #873EFC;
-    border-radius: 5px;
-
-    bottom: -30px;
-    left: 30px;
-}
-
-.layout {
-    display: flex;
-    margin-top: 20px;
-}
-
-.layout>div:nth-child(1) {
-    width: 320px;
-    margin-right: 20px;
-}
-
-.layout>div:nth-child(2) {
-    flex: 1;
-}
-
-.produce-headers {
-    display: flex;
-    font-family: Malgun Gothic;
-    font-size: 14px;
-    border-bottom: 1px solid #E5E5E5;
-}
-
-.produce-header {
-    width: 35px;
-    height: 32px;
-    margin: 10px
-}
-
-.produce-header2 {
-    width: 32px;
-    height: 32px;
-    margin: 10px
-}
-
-.produce-headerline {
-    font-size: 18px;
-    margin-top: 14px;
-}
-
-.pictures {
-    width: 850px;
-    height: 460px;
-    background: white;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    margin-left: 100px;
-    top: 70px;
-    position: sticky;
-}
-
-
-.pictures3 img {
-    width: 175px;
-    height: 175px;
-    margin: 18.4px;
-}
