+++ src/pages/Privatepage.css
@@ -35,7 +35,6 @@ body {
   position: absolute;
   background: #873EFC;
   border-radius: 5px;
-
   bottom: -30px;
   left: 30px;
 }
@@ -110,7 +109,7 @@ body {
   color: #385898;
 }
 
-.pictures {
+.Privatepage-picture>div {
   width: 100%;
   height: 250px;
   background: white;
@@ -121,7 +120,7 @@ body {
   position: sticky;
 }
 
-.pictures3 img {
+.Privatepage-picture>div>p>img {
   width: 100px;
   height: 100px;
   margin: 2.8px;
@@ -233,6 +232,8 @@ body {
 input[type="file"] {
   display: none;
 }
+
+
 
 .custom-file-upload {
   background-color: #F5F6F7;
