+++ src/pages/Profileedit.css
@@ -122,15 +122,5 @@ body {
     margin-left: 27px
 }
 
-.pictures {
-    width: 100%;
-    height: 250px;
-    background: white;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    margin-left: 100px;
-    top: 70px;
-    position: sticky;
-}
 
 /* introdce-container */
